refactor: split go backend modules
